Переработал рендерер.

+++ Core/Source/Graphics/Buffers/IndexBuffer.cpp
@@ -4,12 +4,12 @@ namespace S3DGE
 {
 	namespace Graphics
 	{
-		IndexBuffer::IndexBuffer(ushort* data, uint count)
+		IndexBuffer::IndexBuffer(uint* data, uint count)
 			: m_Count(count)
 		{
 			glGenBuffers(1, &m_BufferID);
 			glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_BufferID);
-			glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(ushort), data, GL_STATIC_DRAW);
+			glBufferData(GL_ELEMENT_ARRAY_BUFFER, count * sizeof(uint), data, GL_STATIC_DRAW);
 			glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, 0);
 		}
