+++ Core/Source/Graphics/Buffers/IndexBuffer.h
@@ -14,7 +14,7 @@ namespace S3DGE
 			uint m_Count;
 
 		public:
-			IndexBuffer(ushort* data, uint count);
+			IndexBuffer(uint* data, uint count);
 			~IndexBuffer();
 
 			inline const uint GetCount() const { return m_Count; }
