+++ Core/Source/Graphics/Renderables/Sprite.h
@@ -10,8 +10,8 @@ namespace S3DGE
 		class Sprite : public Renderable2D
 		{
 		public:
-			Sprite(int x, int y, int width, int height, Maths::vec4f color);
-			Sprite(int x, int y, int width, int height, Texture* texture);
+			Sprite(float x, float y, float width, float height, Maths::vec4f color);
+			Sprite(float x, float y, float width, float height, Texture* texture);
 		};
 	}
 }
